encapsulate script to test
